refactor(http): one copy of RFC 9110's dates, Range and ETag
`HTTP-date`, the `Range` grammar and the `ETag` quoting are facts about
RFC 9110, not about S3, and `mountx/webdav` needs the same three. They move
to `src/http.ts`; `src/s3/protocol.ts` re-exports every symbol under the name
it already had, so `mountx/s3`'s surface is unchanged.

Same argument as the one errno table in `src/errors.ts` (AGENTS.md,
invariant 6): a wire format transcribed twice is a wire format that will be
transcribed differently twice.
